Checking token in pages refactored NOTE: 1. isTokenExist() function in check.token.js file
