+++ frontend/JS/main.js
@@ -1,3 +1,4 @@
+import isTokenExist from "./check.token.js";
 import applyTheme from "./theme.js";
 
 const sampleProducts = [
@@ -61,34 +62,19 @@ document.addEventListener("DOMContentLoaded", async () => {
     updateCartCount();
   }
 
-  // ---------------------------
-  // 🔹 Auth link switcher
-  // ---------------------------
-  const authLink = document.getElementById("auth-link");
-  const token = localStorage.getItem("token");
-
-  if (authLink) {
-    if (token) {
-      authLink.innerHTML = `
-        <a href="#" id="account-btn" class="btn auth-btn">حساب من</a>
-        <a href="#" id="logout-btn" class="btn auth-btn logout-btn">خروج</a>
-      `;
-
-      document.getElementById("logout-btn").addEventListener("click", (e) => {
-        e.preventDefault();
-        localStorage.removeItem("token");
-        localStorage.removeItem("cart");
-        window.location.reload();
-      });
-    } else {
-      authLink.innerHTML = `
-        <a href="login.html" class="btn auth-btn">ورود</a>
-        <a href="signup.html" class="btn auth-btn">ثبت نام</a>
-      `;
-    }
-  }
+  isTokenExist();
 
   applyTheme();
+
+  const logoutBtn = document.getElementById("logout-btn");
+  if (logoutBtn) {
+    logoutBtn.addEventListener("click", (e) => {
+      e.preventDefault();
+      localStorage.removeItem("token");
+      localStorage.removeItem("cart");
+      window.location.reload();
+    });
+  }
 
   const accountBtn = document.getElementById("account-btn");
   if (accountBtn) {
